14.0/extra-addons/pip-requirements.txt
python-pkcs11
lxml
unidecode
num2words
openupgradelib
xlrd
python-dateutil
dicttoxml
PyPDF2
roulier
xmltodict
zeep
cryptography
endesive
py3o.formats
py3o.template
pdf2image
pyzbar
pytz
email-validator
phonenumbers
pycep_correios
pyyaml
satcomum
workalendar
xmldiff
sqlparse>=0.4.2
pandas
openpyxl
imgkit==1.1.0
pyfcm
coverage==5.5
flake8
python-Levenshtein-wheels==0.13.1
erpbrasil.assinatura>=1.7.0
erpbrasil.base>=2.3.0
erpbrasil.edoc
erpbrasil.edoc.pdf
erpbrasil.edoc>=2.5.2
erpbrasil.transmissao
nfelib>=2.0.0
nfselib.ginfes
nfselib.paulistana
mtranslate
signxml
pytrustnfe3
pysigep
iugu
lxml==4.6.3
